+++ NEWS.md
@@ -2,6 +2,7 @@
 **2021-06-10**
 
 - Fix a critical bug in the consolidate and summarize addin
+- Fix problems with chart generation in the analyze addin #46
 
 # phenoptrReports 0.2.11
 **2021-06-01**